more 32 to 16 bit code - change min/maxint used to min/maxlong.. changed a lot of blocklist stuff to 16 bit. demos still seem to run fine.

+++ d_player.h
@@ -162,7 +162,7 @@ typedef struct
     int16_t		skills;
     int16_t		sitems;
     int16_t		ssecret;
-    int32_t		stime; 
+    int16_t		stime; 
   
 } wbplayerstruct_t;
